requirements2.txt
# ============================================
# Smart Revision Generator - Requirements
# Python 3.11
# ============================================

# --- Web Framework ---
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# --- Database ---
supabase==2.4.0
asyncpg==0.29.0

# --- Vector Store ---
faiss-cpu==1.7.4

# --- Embeddings ---
sentence-transformers==2.3.1
torch==2.2.0

# --- Sparse Retrieval ---
rank-bm25==0.2.2

# --- LLM (Google Gemini) ---
google-generativeai==0.4.1

# --- Document Processing ---
PyPDF2==3.0.1
python-docx==1.1.0
tiktoken==0.6.0

# --- Evaluation ---
scikit-learn==1.4.0
rouge-score==0.1.2
numpy==1.26.4

# --- Frontend ---
streamlit==1.31.1

# --- Utilities ---
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.2.1
loguru==0.7.2
tenacity==8.2.3
httpx>=0.24.0

# --- Testing ---
pytest==8.0.2
pytest-asyncio==0.23.5
pytest-cov==4.1.0

# --- Development ---
black==24.2.0
isort==5.13.2
mypy==1.8.0